requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
blinker==1.6.2
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.2
django-celery-results==2.5.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-modeltranslation==0.18.11
django-nine==0.2.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.2
Flask==2.3.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
prompt-toolkit==3.0.38
py==1.11.0
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
retry==0.9.2
send-email==201212
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
